Extract the Subroutines class
